refactor: Refactor markdown2html package structure and add math protection features
- Moved markdown2html from app/http/controllers to app/bundles for better organization.
- Introduced math protection functionality to handle inline and block math segments.
- Added comprehensive tests for markdown to HTML conversion, including math protection and legacy review sections.
- Updated various services and controllers to use the new markdown2html package path.
- Enhanced documentation to reflect changes in the markdown rendering architecture.
