+++ apps/gooseforum/app/models/forum/pageConfig/pageConfig.go
@@ -4,7 +4,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/YourTongji/YourTJ-Hub/apps/gooseforum/app/http/controllers/markdown2html"
+	"github.com/YourTongji/YourTJ-Hub/apps/gooseforum/app/bundles/markdown2html"
 )
 
 const tableName = "page_config"
