+++ apps/gooseforum/app/service/datamigration/reply_markdown.go
@@ -4,7 +4,7 @@ import (
 	"log/slog"
 
 	db "github.com/YourTongji/YourTJ-Hub/apps/gooseforum/app/bundles/connect/dbconnect"
-	"github.com/YourTongji/YourTJ-Hub/apps/gooseforum/app/http/controllers/markdown2html"
+	"github.com/YourTongji/YourTJ-Hub/apps/gooseforum/app/bundles/markdown2html"
 )
 
 type ReplyMarkdownResult struct {
