+++ apps/gooseforum/app/service/postservice/rendered_html.go
@@ -3,7 +3,7 @@ package postservice
 import (
 	"log/slog"
 
-	"github.com/YourTongji/YourTJ-Hub/apps/gooseforum/app/http/controllers/markdown2html"
+	"github.com/YourTongji/YourTJ-Hub/apps/gooseforum/app/bundles/markdown2html"
 	"github.com/YourTongji/YourTJ-Hub/apps/gooseforum/app/models/forum/posts"
 )
